Se agrego el hashTable y el BinaryHeap
